chore: restructure the file tree and code of util functions (#672)
