+++ src/helpers/is-null.ts
@@ -1,0 +1,16 @@
+export function isNull(object: any) {
+  if (
+    object === null ||
+    typeof object === 'undefined' ||
+    object === '' ||
+    JSON.stringify(object) === '[]' ||
+    JSON.stringify(object) === '{}'
+  ) {
+    return true;
+  }
+  return false;
+}
+
+export function isAllNull(obj: Object) {
+  return Object.values(obj).every((value) => value === null);
+}
