+++ src/helpers/is-perceptor.ts
@@ -1,0 +1,4 @@
+const isPerceptor = (): boolean =>
+  window.location.search.includes('?redirect=perceptor');
+
+export default isPerceptor;
