--- src/mock/background.data.ts
@@ -1,23 +1,0 @@
-export const updateInformation = {
-  chrome: {
-    latest_version: '1.8.8',
-    url: 'https://github.com/hypertrons/hypertrons-crx/releases',
-  },
-  edge: {
-    latest_version: '1.8.8',
-    url: 'https://github.com/hypertrons/hypertrons-crx/releases',
-  },
-  develop: {
-    latest_version: '1.8.8',
-    url: 'https://github.com/hypertrons/hypertrons-crx/releases',
-  },
-};
-
-export const notificationInformation = {
-  id: 2021052301,
-  content: {
-    zh: '欢迎使用 Hypertrons-crx !',
-    en: 'Welcome to Hypertrons-crx !',
-  },
-  is_published: true,
-};
